Prefix index: memory usage stats, change default sizing of elements per node to 1, other cleanup/refactor
